LPS-55504 move tests to module
