+++ modules/portal/portal-scripting-javascript/test/unit/com/liferay/portal/scripting/javascript/internal/dependencies/simple.js
